% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multi_correct_function.R
\name{multi_correct}
\alias{multi_correct}
\title{Multilevel Models to Calculate Index of Dissimilarity}
\usage{
multi_correct(group, district1, district2 = 1, comp = 1, data,
  nsim = 1000, seed = 12345)
}
\arguments{
\item{group}{Indicator variable denoting the groups (e.g., 0=male, 1=female or 0=native, 1=immigrant).}

\item{district1}{Units of segregation, denotes the variable that records units of segregation (i.e., districts).}

\item{district2}{Units of segregation in three level model. Defaults to 1 denoting the estimation of a two-level model.}

\item{comp}{Compositional characteristics, list of variables for which D is to be correct for (e.g., age+sex+education). Defaults to 1 denoting that the uncorrected version of D is calculated.}

\item{data}{Name of dataset.}

\item{nsim}{Number of simulations. Defaults to 1000.}

\item{seed}{Seed number. Defaults to 12345.}
}
\description{
This package calculates the Index of Dissimilarity (from hereon denoted as D) using multilevel
logistic regression. Lower level units are nested within the units for which segregation is to be
calculated (e.g., immigrants and natives nested in neighborhoods). The random effects of the
multilevel models serve as "naive" estimator of segregation conceptually and are used to simulated new marginal distributions for each group. The advantage
of using this methods lies in the researchers ability to correct the D for compositional
differences across the two groups by adding corresponding variables as well as correcting for
contextual differences. The function allows the calculation of D via two and three level models.
}
\examples{
## Index of Dissimilarity for immigrants and natives over city districts:
D <- multi_correct(group="immigrants", district1="citydistr", district2=1, comp=1, data=cityofX)
summary(D$Dm1)

## Correcting D for compositional differences
D2 <- multi_correct(group="immigrants", district1="citydistr", district2=1, comp="age+sex+edu+income", data=cityofX)
summary(D2$Dm1)

## D for segregation over cities and their districts
## IMPORTANT: district1 needs to contain unit-ID with the fewest number of units
## we have fewer cities than city districts in the data, therefore city is district1.
D3 <- multi_correct(group="immigrants", district1="city", district2="citydistr", comp="age+sex+edu+income", data=cityofX)

## city segregation
summary(D3$Dm1)

## city district segregation
summary(D3$Dm2)
}
\author{
Christoph Spörlein, \email{christoph.spoerlein@uni-bamberg.de}
}
\references{
Leckie et al. 2012: Multilevel Modeling of Social Segregation. Journal of Educational and Behavioral Statistics 37:3-29
}
\keyword{compositional}
\keyword{differences}
\keyword{dissimilarity}
\keyword{index}
\keyword{multilevel}
\keyword{of}

